+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Removed
+
+- Adopt pm-cli 2026.7.26 typed authoring contracts and remove the any-cast defineExtension shim ([pm-github-zjxx](https://github.com/unbraind/pm-github/blob/main/.agents/pm/tasks/pm-github-zjxx.toon))
+
 ## 2026.7.26 - 2026-07-26
 
 ### Fixed